fix: fixing user data seeding and login
